requirements.txt
streamlit
google-cloud-storage
google-cloud-datastore
google-cloud-pubsub
